tools/snn_fpga_toolkit/fpga-ml-bitstream/requirements.txt
# Core numerical computing
numpy>=1.21.0
scipy>=1.7.0

# Image processing
scikit-image>=0.18.0
Pillow>=8.3.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# Deep Learning (PyTorch)
torch>=1.10.0
torchvision>=0.11.0

# Data handling
pandas>=1.3.0

# CLI utilities
tqdm>=4.62.0
click>=8.0.0

# GUI (PyQt5)
PyQt5>=5.15.0

# Optional: For side-channel analysis
# pyvisa>=1.11.0          # For oscilloscope control
# scikit-learn>=0.24.0     # For additional ML utilities

# Optional: Jupyter notebooks
# jupyter>=1.0.0
# ipywidgets>=7.6.0
